Move beta to root applications directory
