+++ src/applications/beta-enrollment/beta-enrollment-entry.jsx
@@ -1,6 +1,6 @@
-import '../../../platform/polyfills';
+import '../../platform/polyfills';
 
-import startApp from '../../../platform/startup';
+import startApp from '../../platform/startup';
 
 import routes from './routes';
 import manifest from './manifest.json';
