+++ src/applications/beta-enrollment/containers/BetaEnrollmentButton.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { registerBeta } from '../actions';
 import AlertBox from '@department-of-veterans-affairs/formation/AlertBox';
-import { createIsServiceAvailableSelector } from '../../../../platform/user/selectors';
+import { createIsServiceAvailableSelector } from '../../../platform/user/selectors';
 
 class BetaEnrollmentButton extends React.Component {
   static propTypes = {
