+++ src/applications/beta-enrollment/manifest.json
