+++ src/applications/personalization/account-beta/containers/AccountApp.jsx
@@ -7,7 +7,7 @@ import { fetchLatestTerms, acceptTerms } from '../actions';
 import AccountMain from '../components/AccountMain';
 import RequiredLoginView from '../../../../platform/user/authorization/components/RequiredLoginView';
 import DowntimeNotification, { services } from '../../../../platform/monitoring/DowntimeNotification';
-import BetaApp, { features } from '../../beta-enrollment/containers/BetaApp';
+import BetaApp, { features } from '../../../beta-enrollment/containers/BetaApp';
 
 class UserProfileApp extends React.Component {
   render() {
