+++ src/applications/personalization/va-profile-beta/containers/VAProfileApp.jsx
@@ -10,7 +10,7 @@ import {
   clearMessage
 } from '../actions';
 
-import BetaApp, { features } from '../../beta-enrollment/containers/BetaApp';
+import BetaApp, { features } from '../../../beta-enrollment/containers/BetaApp';
 import RequiredLoginView from '../../../../platform/user/authorization/components/RequiredLoginView';
 import DowntimeNotification, { services } from '../../../../platform/monitoring/DowntimeNotification';
 import ProfileView from '../components/ProfileView';
